+++ src/computes/poll.js
@@ -10,14 +10,13 @@
 */
 
 import Constant from '@/data/const';
-import lineType from '@/computes/protocol/line_types';
-import hash from 'js-sha256';
 
 export default class {
-  constructor(pollClient, idbInstance, systemInstance) {
+  constructor(pollClient, instances, system) {
     this.client = pollClient;
-    this.idb = idbInstance;
-    this.system = systemInstance;
+    this.idb = instances.idb;
+    this.listener = instances.operation;
+    this.system = system;
     this.revision = 0;
   }
 
@@ -27,7 +26,9 @@ export default class {
           this.revision,
           Constant.FETCH_OP_NUM,
       );
-      await this.opHandler(operations);
+      for (const operation of operations) {
+        this.listener.trigger(operation);
+      }
       await this.updateRevision(operations);
     } catch (e) {
       console.error(e);
@@ -45,135 +46,5 @@ export default class {
     const latestRev = operations[opLength - 1].revision;
     const secondRev = operations[opLength - 2].revision;
     this.revision = latestRev.compare(secondRev) ? secondRev : latestRev;
-  }
-
-  async updateGroupInfo(groupId, accepted = false) {
-    const data = await this.client.getGroup(groupId);
-    const localData = await this.idb.user.get(
-        Constant.IDB.USER.GROUP.JOINED, data.id,
-    );
-    if (accepted || localData) {
-      this.idb.user.put(Constant.IDB.USER.GROUP.JOINED, data);
-    } else {
-      this.idb.user.put(Constant.IDB.USER.GROUP.INVITED, data);
-      this.system.registerChatRoomIdHash(groupId);
-    }
-  }
-
-  async opHandler(operations) {
-    for (const operation of operations) {
-      switch (operation.type) {
-        case lineType.OpType.UPDATE_PROFILE: {
-          const data = await this.client.getProfile();
-          await this.system.sync.updateProfile(data);
-          break;
-        }
-        case lineType.OpType.ADD_CONTACT:
-        case lineType.OpType.UPDATE_CONTACT: {
-          const data = await this.client.getContact(operation.param1);
-          this.idb.user.put(Constant.IDB.USER.CONTACT, data);
-          break;
-        }
-        case lineType.OpType.ACCEPT_GROUP_INVITATION: {
-          this.idb.user.delete(
-              Constant.IDB.USER.GROUP.INVITED,
-              operation.param1,
-          );
-          await this.updateGroupInfo(operation.param1, true);
-          break;
-        }
-        case lineType.OpType.LEAVE_GROUP: {
-          await this.clearMessageBox(operation.param1);
-          this.vuex.commit(
-              'unregisterChatIdHash',
-              hash.sha256(operation.param1),
-          );
-          this.idb.user.delete(
-              Constant.IDB.USER.GROUP.JOINED,
-              operation.param1,
-          );
-          break;
-        }
-        case lineType.OpType.NOTIFIED_CANCEL_INVITATION_GROUP: {
-          if (operation.param3.includes('\x1e')) {
-            operation.param3 = operation.param3
-                .split('\x1e')
-                .find((id) => id === this.system.profile.userId);
-          }
-          if (operation.param3 === this.system.profile.userId) {
-            this.vuex.commit(
-                'unregisterChatIdHash',
-                hash.sha256(operation.param1),
-            );
-            this.idb.user.delete(
-                Constant.IDB.USER.GROUP.INVITED,
-                operation.param1,
-            );
-          }
-          break;
-        }
-        case lineType.OpType.NOTIFIED_KICKOUT_FROM_GROUP:
-          if (operation.param3.includes('\x1e')) {
-            operation.param3 = operation.param3
-                .split('\x1e')
-                .find((id) => id === this.system.profile.userId);
-          }
-          if (operation.param3 === this.system.profile.userId) {
-            await this.idb.clearMessageBox(operation.param1);
-            this.vuex.commit(
-                'unregisterChatIdHash',
-                hash.sha256(operation.param1),
-            );
-            this.idb.user.delete(
-                Constant.IDB.USER.GROUP.JOINED,
-                operation.param1,
-            );
-          } else {
-            await this.updateGroupInfo(operation.param1);
-          }
-          break;
-        case lineType.OpType.NOTIFIED_UPDATE_GROUP:
-        case lineType.OpType.NOTIFIED_INVITE_INTO_GROUP:
-        case lineType.OpType.NOTIFIED_ACCEPT_GROUP_INVITATION:
-        case lineType.OpType.NOTIFIED_LEAVE_GROUP:
-        case lineType.OpType.CREATE_GROUP:
-        case lineType.OpType.UPDATE_GROUP:
-        case lineType.OpType.CANCEL_INVITATION_GROUP:
-        case lineType.OpType.INVITE_INTO_GROUP:
-        case lineType.OpType.KICKOUT_FROM_GROUP:
-          await this.updateGroupInfo(operation.param1);
-          break;
-        case lineType.OpType.SEND_MESSAGE:
-        case lineType.OpType.RECEIVE_MESSAGE:
-          // Add Target for index
-          operation.message.target = (() => {
-            switch (operation.message.toType) {
-              case lineType.MIDType.USER:
-                if (operation.message.from_ === this.system.profile.userId) {
-                  return operation.message.to;
-                } else {
-                  return operation.message.from_;
-                }
-              case lineType.MIDType.ROOM:
-              case lineType.MIDType.GROUP:
-                return operation.message.to;
-            }
-          })();
-          // Uint8Array to String
-          operation.message.createdTime =
-                        operation.message.createdTime.toString();
-          operation.message.deliveredTime =
-                        operation.message.deliveredTime.toString();
-          this.idb.user.put(
-              Constant.IDB.USER.PREVIEW_MESSAGE_BOX,
-              operation.message,
-          );
-          this.idb.user.put(
-              Constant.IDB.USER.MESSAGE_BOX,
-              operation.message,
-          );
-          break;
-      }
-    }
   }
 }
